+++ CHANGES.rst
@@ -3,6 +3,7 @@ Version 2.1.3
 
 Unreleased
 
+-   Allow Sphinx's parallel build feature. :issue:`88`
 
 Version 2.1.2
 -------------